// src/type.ts
import type { Request } from 'express';
export type RedirectFn = (data: {
  req: Request;
  redirect_uri: string;
  state: string;
}) => Promise<{ redirectUrl: string }>;

export type GetUserInfoFn = (code: string) => Promise<{
  username: string;
  avatar: string;
  contact: string;
  memberName?: string;
}>;

export type CallbackFn = (data: { req: Request }) => Promise<{ redirectUrl: string }>;

export type AssertFn = (data: {
  SAMLResponse: string;
  RelayState: string;
}) => Promise<{ redirectUrl: string }>;
export type GetMetaDataFn = () => Promise<string>;

export type UserListType = {
  username: string;
  memberName: string;
  avatar?: string;
  contact?: string;
  orgs?: string[];
}[];

export type GetUserListFn = () => Promise<UserListType>;

export type OrgListType = {
  id: string;
  name: string;
  parentId: string;
}[];

export type GetOrgListFn = () => Promise<OrgListType>;

export type lcfcUser = {
  name: string;
  acctName: string;
  key: string;
  status: string;
  isPublic: string;
  isPartners: string;
  period: string;
  createTime: string;
  disableTime: string;
};

export type LcfcUserListType = lcfcUser[];